requirements.txt
langchain-chroma==0.1.4
langchain_openai==0.2.1
langchain==0.3.1
Pillow==10.4.0